InteractiveTradeR

API for trading, data collection, and strategy testing.

Disclaimer and Waiver

You can use this package to make actual trades on real money if you connect it to one of your live accounts.

Not affiliated with Interactive Brokers.

Only a fool would place blind trust into some random package downloaded off the Internet.

Package options

InteractiveTradeR uses the following [options()] to configure your connections to Trader Workstation (TWS) and IB Gateway (IBG):

`interactivetrader.paper`: Set to TRUE to use options configured for **PAPER** account; set to FALSE for **LIVE**. Defaults to TRUE. **Using InteractiveTradeR to trade real money without a license violates the license agreement**.

`interactivetrader.platform`: Set to "TWS" to use options configured for **Trader Workstation** account; set to "IBG" for **IB Gateway**. Defaults to TWS.

#'

`interactivetrader.tws.paper.host`: Identifies its set value as the host name at which InteractiveTradeR should look to connect to an instance of Trader Workstation connected to a **PAPER** account. Because most users run TWS on the same machine as InteractiveTradeR, this option should usually be kept to the default setting of "localhost".

`interactivetrader.tws.paper.host`: Identifies its set value as the host name at which InteractiveTradeR should look to connect to an instance of Trader Workstation connected to a **PAPER** account. Because most users run TWS on the same machine as InteractiveTradeR, this option should usually be kept to the default setting of "localhost".

`interactivetrader.tws.paper.port`: Identifies the port on which an instance of Trader Workstation, connected to a **PAPER** account, has been configured to listen for API connections. Default is **7497**, matching the default set by Interactive Brokers.

`interactivetrader.tws.paper.master`: To use a master_client_idMaster Client ID that you have configred within TWS for a **PAPER** account, set this option to the value you have selected for your TWS paper account's Master Client ID.

`interactivetrader.ibg.paper.host`: Identifies its set value as the host name at which InteractiveTradeR should look to connect to an instance of IB Gateway connected to a **PAPER** account. Because most users run IBG on the same machine as InteractiveTradeR, this option should usually be kept to the default setting of "localhost".

`interactivetrader.ibg.paper.port`: Identifies the port on which an instance of IB Gateway, connected to a **PAPER** account, has been configured to listen for API connections. Default is **4002**, matching the default set by Interactive Brokers.

`interactivetrader.ibg.paper.master`: To use a master_client_idMaster Client ID that you have configred within IBG for a **PAPER** account, set this option to the value you have selected for your IBG paper account's Master Client ID.

`interactivetrader.tws.live.host`: Identifies its set value as the host name at which InteractiveTradeR should look to connect to an instance of Trader Workstation connected to a **LIVE** account on which actual trades will be executed with real money. Because most users run TWS on the same machine as InteractiveTradeR, this option should usually be kept to the default setting of "localhost". **Using InteractiveTradeR to trade real money without a license violates the license agreement**.

`interactivetrader.tws.live.port`: Identifies the port on which an instance of Trader Workstation, connected to a **LIVE** account, has been configured to listen for API connections. Default is **7496**, matching the default set by Interactive Brokers. **Using InteractiveTradeR to trade real money without a license violates the license agreement**.

`interactivetrader.tws.live.master`: To use a master_client_idMaster Client ID that you have configred within TWS for a **LIVE** account, set this option to the value you have selected for your TWS paper account's Master Client ID. **Using InteractiveTradeR to trade real money without a license violates the license agreement**.

`interactivetrader.ibg.live.host`: Identifies its set value as the host name at which InteractiveTradeR should look to connect to an instance of IB Gateway connected to a **LIVE** account on which actual trades will be executed with real money. Because most users run IBG on the same machine as InteractiveTradeR, this option should usually be kept to the default setting of "localhost". **Using InteractiveTradeR to trade real money without a license violates the license agreement**.

`interactivetrader.ibg.live.port`: Identifies the port on which an instance of IB Gateway, connected to a **LIVE** account, has been configured to listen for API connections. Default is **4001**, matching the default set by Interactive Brokers. **Using InteractiveTradeR to trade real money without a license violates the license agreement**.

`interactivetrader.ibg.live.master`: To use a master_client_idMaster Client ID that you have configred within IBG for a **LIVE** account, set this option to the value you have selected for your IBG paper account's Master Client ID. **Using InteractiveTradeR to trade real money without a license violates the license agreement**.