Order Objects

* **orderComboLegs** <numeric>: A numeric vector whose length equals the number of rows in the `comboLegs` element in the contract object. `orderComboLegs` defines the order prices for each combo leg.

describe

Details

* **smartComboRoutingParams** <named character>: In InteractiveTradeR, `smartComboRoutingParams` is a named character vector in which a parameter is identified by a "tag", and its value by "value". The tags and values available are described in the smartComboRoutingParams section of IB's documentation.

For example, to specify a non-gauranteed order, use `smartComboRoutingParams = c(NonGuaranteed = 1)`.

To instruct the execution system to not start the next leg-in if the previous leg-in did not finish, you would use `smartComboRoutingParams = c(DontLeginNext = 1)`

You may specify as many smart combo routing params as you like; simply use `c`() to combine them into a named character vector; for example: `smartComboRoutingParams = c(NonGuaranteed = 1, DontLeginNext = 1)`