Place Order

Sends an order to IB's servers to buy or sell a financial instrument.

place_order(
  contract,
  order,
  deltaNeutralContract = NULL,
  softDollarTier = NULL,
  channel = NULL,
  non_master_override = TRUE
)

Arguments

contract
Named character vector of contract parameter(s). The name of each element identifies the parameter -- for example, secType, symbol, exchange... -- and each element itself gives the set value (e.g., "STK", "IBM", "SMART").

The `contract` parameters that may be used with place_order() are:
deltaNeutralContract, secId, conId, comboLegs, secType, tradingClass, symbol, lastTradeDateOrContractMonth, strike, right, multiplier, exchange, primaryExchange, currency, localSymbol, secIdType

See contract for detailed information on all contract parameters.

order
describe that

deltaNeutralContract
that too

softDollarTier
and that

channel
One of the following:

Not Specified (Default): Opens a new connection to IB, uses it to issue the request and retrieve the response, and closes connection behind itself upon completion.

The Name of a Sock: Character vector, length 1. The name of an open, connected socket in the sock_drawer; e.g., "master", "tws", or "sock_123"

Numeric Client ID: Numeric, length 1. The client ID for which open orders are to be retrieved; e.g., 0, 874, 123. If a client ID is passed, and no socket in the sock_drawer is connected on that ID, then a new socket will be opened on that ID, and closed upon function exit.

A sockconn Connection: An open connection object of class "sockconn", connected to the IB API; e.g., sock_drawer$tws

non_master_override
finally that

Conditional Orders

Conditional Orders

Bracket Orders

Bracket Orders

Attaching (Hedging) Orders

Beta Hedge:

Delta Hedge:

FX Order:

Pair Trade:

Combo Limit:

Combo Market:

Examples

# 1a) Basic Limit Order for a stock
place_order(
  contract = c(
    symbol   = "IBKR",
    secType  = "STK",
    currency = "USD",
    exchange = "SMART"
  ),
  order    = c(
    action        = "BUY",
    orderType     = "LMT",
    totalQuantity = 1,
    lmtPrice      = 45,
    account       = req_account_summary(tags = "AvailableFunds") %>%
      dplyr::filter(as.numeric(tag_value) > 5000) %>%
      dplyr::select(account) %>%
      unlist(use.names = FALSE) %>%
      sample(1)
  )
)

#> [1] "yolo"

# 1b) Basic Limit Order for a stock's lowest-strike in-the-money call option
place_order(
  contract = c(
    symbol   = "IBKR",
    secType  = "STK",
    currency = "USD",
    exchange = "SMART"
  ),
  order    = c(
    action        = "BUY",
    orderType     = "LMT",
    totalQuantity = 1,
    lmtPrice      = 45,
    account       = req_account_summary(tags = "AvailableFunds") %>%
      dplyr::filter(as.numeric(tag_value) > 5000) %>%
      dplyr::select(account) %>%
      unlist(use.names = FALSE) %>%
      sample(1)
  )
)

#> [1] "yolo"

# 2) Market Combo Order
place_order(
  contract = list(
    symbol    = "IBKR,MCD",
    secType   = "BAG",
    currency  = "USD",
    exchange  = "SMART",
    comboLegs = tibble::tibble(
      conId    = c(43645865, 9408),
      ratio    = c(1,1),
      action   = c("BUY","SELL"),
      exchange = c("SMART", "SMART")
    )
  ),
  order = list(
    action        = "BUY",
    orderType     = "MKT",
    totalQuantity = 1,
    account       = req_account_summary(tags = "AvailableFunds") %>%
      dplyr::filter(as.numeric(tag_value) > 5000) %>%
      dplyr::select(account) %>%
      unlist(use.names = FALSE) %>%
      sample(1),
    smartComboRoutingParams = c(NonGuaranteed = 1)
  )
)

#> [1] "yolo"

### 3) Limit Combo Order, Combo Price Specified
place_order(
  contract = list(
    symbol    = "IBKR,MCD",
    secType   = "BAG",
    currency  = "USD",
    exchange  = "SMART",
    comboLegs = tibble::tibble(
      conId    = c(43645865, 9408),
      ratio    = c(1,1),
      action   = c("BUY","SELL"),
      exchange = c("SMART", "SMART")
    )
  ),
  order = list(
    action        = "BUY",
    orderType     = "LMT",
    lmtPrice      = 150,
    totalQuantity = 1,
    account       = req_account_summary(tags = "AvailableFunds") %>%
      dplyr::filter(as.numeric(tag_value) > 5000) %>%
      dplyr::select(account) %>%
      unlist(use.names = FALSE) %>%
      sample(1),
    smartComboRoutingParams = c(NonGuaranteed = 1)
  )
)

#> [1] "yolo"

### 4) Limit Combo Order, Individual Leg Prices Specified
place_order(
  contract = list(
    symbol    = "IBKR,MCD",
    secType   = "BAG",
    currency  = "USD",
    exchange  = "SMART",
    comboLegs = tibble::tibble(
      conId    = c(43645865, 9408),
      ratio    = c(1,1),
      action   = c("BUY","SELL"),
      exchange = c("SMART", "SMART")
    )
  ),
  order = list(
    action         = "BUY",
    orderType      = "LMT",
    orderComboLegs = c(45,210),
    totalQuantity  = 1,
    account        = req_account_summary(tags = "AvailableFunds") %>%
      dplyr::filter(as.numeric(tag_value) > 5000) %>%
      dplyr::select(account) %>%
      unlist(use.names = FALSE) %>%
      sample(1),
    smartComboRoutingParams = c(NonGuaranteed = 1)
  )
)

#> [1] "yolo"

### 5) Delta Neutral Order
place_order(
  contract = list(
    symbol     = "GOOG",
    secType    = "OPT",
    currency   = "USD",
    exchange   = "SMART",
    right      = "C",
    multiplier = 100,
    strike     = 1170
  ),
  order = list(
    action         = "BUY",
    orderType      = "MKT",
    totalQuantity  = 1,
    account        = req_account_summary(tags = "AvailableFunds") %>%
      dplyr::filter(as.numeric(tag_value) > 5000) %>%
      dplyr::select(account) %>%
      unlist(use.names = FALSE) %>%
      sample(1)
  )
)

#> [1] "yolo"