Request Account Updates Multi

This function either fetches or sets up a subscription to receive the data that appear in the Account Window window of Trader Workstation. req_account_updates_multi() may be used to set up multiple simultaneous subscriptions to many accounts / model codes. IB's documentation may be found on the Account Updates page.

req_account_updates_multi(
  account = "All",
  modelCode = NULL,
  ledgerAndNLV = FALSE,
  channel = NULL,
  return_data = is.null(channel)
)

Arguments

account
Character vector of length 1 containing a valid account code.

modelCode
Character vector of length 1 containing a valid model code.

ledgerAndNLV
Boolean of length 1, defaults to FALSE. If TRUE, then the created subscriptions are treated as "lightweight requests", meaning that the data returned for the account or model codes supplied will include only currency positions (as opposed to both account values and currency positions).

channel
One of the following:

Not Specified (Default): Opens a new connection to IB, uses it to issue the request and retrieve the response, and closes connection behind itself upon completion.

The Name of a Sock: Character vector, length 1. The name of an open, connected socket in the sock_drawer; e.g., "master", "tws", or "sock_123"

Numeric Client ID: Numeric, length 1. The client ID for which open orders are to be retrieved; e.g., 0, 874, 123. If a client ID is passed, and no socket in the sock_drawer is connected on that ID, then a new socket will be opened on that ID, and closed upon function exit.

A sockconn Connection: An open connection object of class "sockconn", connected to the IB API; e.g., sock_drawer$tws

return_data
Boolean of length 1. Defaults to TRUE unless argument channel is specified. If FALSE, data retrieved by the function will be returned as the funciton's output. If TRUE, then a Boolean succeses flag will be returned as the function's output indicating the success (TRUE) or failure (FALSE) of the function's attempt to transceive data to/from IB. Data in the treasury is always updated regardless of the value passed as return_data in the function call.

Value

This function is called for its side effect of updating the treasury, which takes place every time the function executes. Additionally, the function's return value depends upon the value passed in as return_data as follows:

If return_data == FALSE: A Boolean success flag, returned invisibly, indicating that the function executed correctly and updated the treasury with any new data retrieved.

If return_data == TRUE: Any new data retrieved will be returned in a tibble in addition to being added to the treasury. If no new data is available, returns NULL.

return_data defaults to TRUE unless channel is specified.

Details

"All" Option: If account is not specified, req_account_updates_multi() will default to account = "All", which will return data for all accounts accessible by the signed-in user. This option is not available for users who manage more than 50 accounts.

Models: Model Portfolios, or "Models" for short, can be though of as named templates that will allocate funds across investment products in a specified manner. When passed to req_account_updates_multi() as model_code, they behave like stocks -- they have postions, market prices, values, etc. To use Model Portfolios you need to have a Financial Advisor account, and the Model Portfolios feature needs to be activated. Learn more on Interactive Brokers' Model Portfolios page.

No Positions: If you call req_account_updates_multi() on an account that has cash but no positions, then data will be sent to the socket once initially, but will not update thereafter.

ACCOUNTS Treasury Object

req_account_summary() updates the ACCOUNTS object in the treasury. ACCOUNTS is a tibble in which each row represents a parameter pertaining to a particular account. Has the following columns:

account <chr>: Account ID (e.g., "DF1234567")

tag <chr>: Name of account tag (e.g., "BuyingPower")

tag_value <chr>: Value of tag (e.g., "500000")

currency <chr>: 3-letter currency abbreviation if tag_value is a monetary amount, "" otherwise.

See also