Request Managed Accounts

Retrieve the account ID codes for all IB accounts your username can access.

req_managed_accts(channel = NULL)

Arguments

channel
One of the following:

Not Specified (Default): Opens a new connection to IB, uses it to issue the request and retrieve the response, and closes connection behind itself upon completion.

The Name of a Sock: Character vector, length 1. The name of an open, connected socket in the sock_drawer; e.g., "master", "tws", or "sock_123"

Numeric Client ID: Numeric, length 1. The client ID for which open orders are to be retrieved; e.g., 0, 874, 123. If a client ID is passed, and no socket in the sock_drawer is connected on that ID, then a new socket will be opened on that ID, and closed upon function exit.

A sockconn Connection: An open connection object of class "sockconn", connected to the IB API; e.g., sock_drawer$tws

Value

character vector in which each element is the ID of an account accessible by the currently logged-in user.

See also

Examples

# Fetch managed accounts
managed_accts <- req_managed_accts()
# Print managed accounts
managed_accts

#> [1] "DF1267859" "DU1267860" "DU1267861" "DU1267862" "DU1267863" "DU1267864"