Request Market Data

Fetch live (or 15-minute delayed) market data.

req_mkt_data(
  contract,
  data_name,
  mktDataType = "REALTIME",
  genericTickList = NULL,
  snapshot = TRUE,
  regulatorySnapshot = FALSE,
  channel = NULL
)

Arguments

contract
Named character vector of contract parameter(s). The name of each element identifies the parameter -- for example, secType, symbol, exchange... -- and each element itself gives the set value (e.g., "STK", "IBM", "SMART").

The `contract` parameters that may be used with place_order() are:
deltaNeutralContract, secId, conId, comboLegs, secType, tradingClass, symbol, lastTradeDateOrContractMonth, strike, right, multiplier, exchange, primaryExchange, currency, localSymbol, secIdType

See contract for detailed information on all contract parameters.

data_name
Character vector of length 1. An identifier for the market data subscription. Good choices could be the contract's symbol, conId, or description (if combo). All market data received by a mkt_data subscription will be stored in the mkt_data environment under the name data_name. Therefore, no two mkt_data subscriptions may have identical data_names.

mktDataType
Either:

String having one of the following values: "REALTIME", "FROZEN", "DELAYED", or "DELAYED_FROZEN".

Integer from 1 to 4, inclusive, corresponding respectively to each of the above choices (e.g., mktDataType = 3 will set the market data type to "DELAYED").

genericTickList
moar biz

snapshot
some biz

regulatorySnapshot
some other biz

channel
One of the following:

Not Specified (Default): Opens a new connection to IB, uses it to issue the request and retrieve the response, and closes connection behind itself upon completion.

The Name of a Sock: Character vector, length 1. The name of an open, connected socket in the sock_drawer; e.g., "master", "tws", or "sock_123"

Numeric Client ID: Numeric, length 1. The client ID for which open orders are to be retrieved; e.g., 0, 874, 123. If a client ID is passed, and no socket in the sock_drawer is connected on that ID, then a new socket will be opened on that ID, and closed upon function exit.

A sockconn Connection: An open connection object of class "sockconn", connected to the IB API; e.g., sock_drawer$tws

Details

attributes `canAutoExecute`, `pastLimit`, `preOpen` are unnecessary in InteractiveTradeR because you can easily create your own flags that accept the `trading_floor$pit` data as input. The aren't that much use anyway, and are mostly holdovers from earlier versions & systems.

See also